change task in homework dz7
