+++ src/main/java/homework_dz7/hw_task2_1234.java
@@ -3,7 +3,7 @@ package homework_dz7;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class hw_task2 {
+public class hw_task2_1234 {
     public static void main(String[] args) {
         String str = "Here you can find activities to practise your reading skills. Reading will help you to improve your understanding of the language and build your vocabulary.";
         String UpperCaseStr = str.toUpperCase();
@@ -42,38 +42,5 @@ public class hw_task2 {
         System.out.println(str.replaceAll("a", "@"));
         System.out.println(str.replaceAll("c", " "));
 
-        int letters = 0;
-        for (Character letter : str.toCharArray()) {
-            letters = letters +1;
-        }
-        System.out.println("Character count = " + letters);
-        System.out.println("");
-
-//        Pattern pattern = new Pattern("\\d+");
-//        String code = "AX6BYU56UX6CV6BNT7NM";
-//        Matcher matcher = pattern.matcher(code);
-//        int start = 0;
-//        while (matcher.find(start)) {
-//            String numbers = code.substring(matcher.start(), matcher.end());
-//            int res = Integer.parseInt(numbers);
-//            System.out.println(res);
-//            start = matcher.end();
-//        }
-        Pattern pat=Pattern.compile("(\\d)\\D*(\\d)");
-        Matcher matcher=pat.matcher("AX6BYU56UX6CV6BNT7NM");
-        int digits = 1;
-        while (matcher.find()) {
-            System.out.println("numbers code: " + matcher.group(1) + matcher.group(2));
-            digits = digits * Integer.parseInt(matcher.group(1) + matcher.group(2));
-        };
-        System.out.println("AX6BYU56UX6CV6BNT7NM" + " " + digits);
-        System.out.println("");
-
-
-        char [] Text = {'h', 'e', 'l', 'p'};
-        System.out.println(String.copyValueOf(Text, 0, 4));
-        System.out.println(String.copyValueOf(Text, 1, 3) + String.copyValueOf(Text, 0, 1));
-        System.out.println(String.copyValueOf(Text, 2, 2) + String.copyValueOf(Text, 0, 2));
-        System.out.println(String.copyValueOf(Text, 3, 1) + String.copyValueOf(Text, 0, 3));
     }
 }
